Move legacy running queries to legacyqueryrunner
